+++ jest.config.ts
@@ -15,6 +15,8 @@ const config: Config.InitialOptions = {
       "<rootDir>/packages/implementations/react/src/atomic/Box/src",
     "@rarui-react/button":
       "<rootDir>/packages/implementations/react/src/atomic/Button/src",
+    "@rarui-react/icon":
+      "<rootDir>/packages/implementations/react/src/atomic/Icon/src",
     "@rarui-react/link":
       "<rootDir>/packages/implementations/react/src/atomic/Link/src",
     "@rarui-react/text":