+++ packages/implementations/react/src/atomic/Toast/src/components/ToastContainer.types.ts
@@ -1,0 +1,17 @@
+import { OlHTMLAttributes } from "react";
+
+export interface ToastContainer {
+  /**
+   * The position of the container in the page
+   * @Defautl bottomRight
+   */
+  position?: "bottomRight" | "bottomLeft" | "topRight" | "topLeft";
+  /**
+   * The size of the toast
+   * @Default medium
+   */
+  size?: "small" | "medium";
+}
+
+export type ToastContainerProps = ToastContainer &
+  OlHTMLAttributes<HTMLOListElement>;