+++ packages/implementations/react/src/atomic/Toast/src/index.ts
@@ -1,0 +1,5 @@
+import { Toast } from "./Toast";
+
+export { Toast } from "./Toast";
+export type { ToastProps } from "./toast.types";
+export default Toast;