+++ packages/implementations/react/src/index.ts
@@ -10,4 +10,5 @@ export * from "@rarui-react/skeleton";
 export * from "@rarui-react/status";
 export * from "@rarui-react/text";
 export * from "@rarui-react/title";
+export * from "@rarui-react/toast";
 export * from "@rarui-react/toggle";