+++ packages/styles/src/index.ts
@@ -13,6 +13,7 @@ export { skeleton } from "./packages/atomic/skeleton";
 export { status } from "./packages/atomic/status";
 export { text } from "./packages/atomic/text";
 export { title } from "./packages/atomic/title";
+export { toast } from "./packages/atomic/toast";
 export { toggle } from "./packages/atomic/toggle";
 
 // types
@@ -27,4 +28,5 @@ export type { ButtonVariants } from "./packages/atomic/button";
 export type { CheckboxVariants } from "./packages/atomic/checkbox";
 export type { ChipVariants } from "./packages/atomic/chip";
 export type { StatusVariants } from "./packages/atomic/status";
+export type { ToastVariants } from "./packages/atomic/toast";
 export type { ToggleVariants } from "./packages/atomic/toggle";