+++ packages/styles/src/packages/atomic/toast/index.ts
@@ -1,0 +1,7 @@
+import { styles } from "./rarui-toast.css";
+
+export const toast = {
+  classnames: { ...styles },
+};
+
+export type { ToastVariants } from "./rarui-toast.types";